:truck: move: 스타일 파일 styles directory로 이동
