+++ grass-diary/src/pages/Setting/Temporary.tsx
@@ -1,4 +1,4 @@
-import * as S from './styles';
+import * as S from '../../styles/Setting/SettingStyles';
 
 const Temporary = () => {
   return (
